+++ packages/dart_http_server_runtime/lib/src/native/http_native_bridge.dart
@@ -1,5 +1,5 @@
 export 'generated_bindings.dart' show NativeBytes, NativeOwnedBytes, NativePair;
-export 'native_binary_payload_lease.dart' show NativeBinaryPayloadLease;
+export 'native_binary_payload_lease.dart' show RuntimeNativeBinaryPayloadLease;
 export 'native_value_helpers.dart'
     show
         NativeAllocations,